Refactoring javascript. Better way to generate graphs.
