+++ gui-flask/static/js-qwak/js-dynamic/dynamicQuantumwalk.js
