+++ gui-flask/static/js-qwak/js-static/static-tools.js
