+++ gui-flask/static/js-qwak/js-static/staticQuantumwalk.js
