+++ gui-flask/templates/dynamicQW.html
@@ -30,6 +30,7 @@
     <script type='module' src="{{ url_for('static',filename='js-qwak/dynamic-script.js') }}"></script>
     <script type='module' src="{{ url_for('static',filename='js-qwak/js-dynamic/dynamic-graph.js') }}"></script>
     <script type='module' src="{{ url_for('static',filename='js-qwak/js-dynamic/dynamic-probDist.js') }}"></script>
+    <script type='module' src="{{ url_for('static',filename='js-qwak/js-dynamic/dynamic-tools.js') }}"></script>
 </head>
 <body class="is-preload">
 
@@ -40,9 +41,8 @@
         <nav>
             <ul>
                 <li><a href="{{url_for('home')}}">Home</a></li>
-                <li><a href="{{url_for('staticQW')}}" class="active">Static</a></li>
-                <li><a href="{{url_for('dynamicQW')}}">Dynamic</a></li>
-<!--                <li><a href="elements.html">Elements</a></li>-->
+                <li><a href="{{url_for('staticQW')}}">Static</a></li>
+                <li><a href="{{url_for('dynamicQW')}}" class="active">Dynamic</a></li>
             </ul>
         </nav>
     </header>
@@ -62,12 +62,11 @@
                                 <button id="defaultOpen" class="tablinks">Graph Generator</button>
                                 <button id="customGraphDisplay" class="tablinks">Custom Graph</button>
                             </div>
-                            <script type='module' src="{{ url_for('static',filename='js-qwak/dynamic-tools.js') }}"></script>
                             <!-- Tab content -->
                             <div id="GraphGenerator" class="tabcontent">
                                 <div class='set-graph-combo'>
                                     <div class="graph-container">
-                                        <div id="cy" class="graph-display"></div>
+                                        <div id="cyDynamic" class="graph-display"></div>
                                     </div>
                                     <button id="runGraphButton" class="button fit">Set Graph</button>
                                     <li class='input-combo'>
@@ -200,21 +199,16 @@
     </footer>
 </div>
 <!-- Scripts -->
-           <script src="{{url_for('static', filename='js/jquery.min.js')}}"></script>
-        <script src="{{url_for('static', filename='js/jquery.scrollex.min.js')}}"></script>
-        <script src="{{url_for('static', filename='js/jquery.scrolly.min.js')}}"></script>
-        <script src="{{url_for('static', filename='js/browser.min.js')}}"></script>
-        <script src="{{url_for('static', filename='js/breakpoints.min.js')}}"></script>
-        <script src="{{url_for('static', filename='js/util.js')}}"></script>
-        <script src="{{url_for('static', filename='js/main.js')}}"></script>
-
+<script src="{{url_for('static', filename='js/jquery.min.js')}}"></script>
+<script src="{{url_for('static', filename='js/jquery.scrollex.min.js')}}"></script>
+<script src="{{url_for('static', filename='js/jquery.scrolly.min.js')}}"></script>
+<script src="{{url_for('static', filename='js/browser.min.js')}}"></script>
+<script src="{{url_for('static', filename='js/breakpoints.min.js')}}"></script>
+<script src="{{url_for('static', filename='js/util.js')}}"></script>
+<script src="{{url_for('static', filename='js/main.js')}}"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/mathjs/10.1.1/math.min.js"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/lodash.js/4.17.10/lodash.js"></script>
 <script src="https://cdn.jsdelivr.net/npm/cytoscape-edgehandles@4.0.1/cytoscape-edgehandles.min.js"></script>
-<!-- <script src="https://unpkg.com/cytoscape-popper@2.0.0/cytoscape-popper.js"></script> -->
-<!--<script src="https://code.jquery.com/jquery-3.2.1.slim.min.js"-->
-<!--        integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN"-->
-<!--        crossorigin="anonymous"></script>-->
 <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.11.0/umd/popper.min.js"
         integrity="sha384-b/U6ypiBEHpOf/4+1nzFpr53nxSS+GLCkfwBdFNTxtclqqenISfwAzpKaMNFNmj4"
         crossorigin="anonymous"></script>
