+++ gui-flask/templates/staticQW.html
@@ -30,6 +30,8 @@
     <script type='module' src="{{ url_for('static',filename='js-qwak/static-script.js') }}"></script>
     <script type='module' src="{{ url_for('static',filename='js-qwak/js-static/static-graph.js') }}"></script>
     <script type='module' src="{{ url_for('static',filename='js-qwak/js-static/static-probDist.js') }}"></script>
+    <script type='module' src="{{ url_for('static',filename='js-qwak/js-static/static-tools.js') }}"></script>
+
 </head>
 
 <body class="is-preload static">
@@ -61,7 +63,6 @@
                                 <button id="customGraphDisplay" class="tablinks">Custom Graph</button>
                             </div>
                             <!--Tab links-->
-                            <script type='module' src="{{ url_for('static',filename='js-qwak/static-tools.js') }}"></script>
                             <!-- Tab content -->
                             <div id="GraphGenerator" class="tabcontent">
                                 <!--                            <div class='grid-element'>-->
@@ -209,21 +210,16 @@
 </div>
 
 <!-- Scripts -->
-        <script src="{{url_for('static', filename='js/jquery.min.js')}}"></script>
-        <script src="{{url_for('static', filename='js/jquery.scrollex.min.js')}}"></script>
-        <script src="{{url_for('static', filename='js/jquery.scrolly.min.js')}}"></script>
-        <script src="{{url_for('static', filename='js/browser.min.js')}}"></script>
-        <script src="{{url_for('static', filename='js/breakpoints.min.js')}}"></script>
-        <script src="{{url_for('static', filename='js/util.js')}}"></script>
-        <script src="{{url_for('static', filename='js/main.js')}}"></script>
-
+<script src="{{url_for('static', filename='js/jquery.min.js')}}"></script>
+<script src="{{url_for('static', filename='js/jquery.scrollex.min.js')}}"></script>
+<script src="{{url_for('static', filename='js/jquery.scrolly.min.js')}}"></script>
+<script src="{{url_for('static', filename='js/browser.min.js')}}"></script>
+<script src="{{url_for('static', filename='js/breakpoints.min.js')}}"></script>
+<script src="{{url_for('static', filename='js/util.js')}}"></script>
+<script src="{{url_for('static', filename='js/main.js')}}"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/mathjs/10.1.1/math.min.js"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/lodash.js/4.17.10/lodash.js"></script>
 <script src="https://cdn.jsdelivr.net/npm/cytoscape-edgehandles@4.0.1/cytoscape-edgehandles.min.js"></script>
-<!-- <script src="https://unpkg.com/cytoscape-popper@2.0.0/cytoscape-popper.js"></script> -->
-<!--<script src="https://code.jquery.com/jquery-3.2.1.slim.min.js"-->
-<!--        integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN"-->
-<!--        crossorigin="anonymous"></script>-->
 <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.11.0/umd/popper.min.js"
         integrity="sha384-b/U6ypiBEHpOf/4+1nzFpr53nxSS+GLCkfwBdFNTxtclqqenISfwAzpKaMNFNmj4"
         crossorigin="anonymous"></script>
